@@ -1,4 +1,5 @@
-module.exports = {
+import BotCommand from '../../types/command/BotCommand.js'
+export default new BotCommand({
   name: 'blank',
   usage: 'blank',
   aliases: ['blank'],
@@ -9,4 +10,4 @@ module.exports = {
   args: false,
   run: function(msg, args) {
   }
-}
+})
