@@ -1,4 +1,5 @@
-module.exports = {
+import BotCommand from '../../types/command/BotCommand.js'
+export default new BotCommand({
     name: 'storedinfo',
     usage: 'storedinfo <@user>',
     aliases: [],
@@ -10,10 +11,10 @@ module.exports = {
     run: function(msg, args) {
         const collection = msg.client.database.collection('users')
         const userID = args[0].replace(/\D/g, '')
-        msg.client.fetchUser(userID, false).then(info => {
+        msg.client.users.fetch(userID, false).then(info => {
             collection.findOne({ userID })
           .then(data => msg.channel.sendMsgEmbed('```json\n' + JSON.stringify(data) + '```'))
           .catch(e => msg.channel.sendMsgEmbed(`There was an error with retrieving the data.`))
         })
     }
-}
+})
