@@ -1,12 +1,23 @@
-const TestBot = require('./classes/TestBot')
+import TestBot from './classes/TestBot.js'
 
-module.exports = async client => {
+// import test suites
+import commandTest from './cases/commands.test.js'
+import apiTest from './cases/api.test.js'
+import gameTest from './cases/games.test.js'
+import DummyAccount from './classes/DummyAccount.js'
+
+export default async client => {
     // Initialize
-    const tester = new TestBot(process.env.DISCORD_TEST_BOT_TOKEN, client, process.env.TEST_CHANNEL)
+    const dummyTokens = process.env.DISCORD_TEST_DUMMY_TOKENS.split(',')
+    const channel = process.env.TEST_CHANNEL
+    const dummyBots = [
+        new DummyAccount(dummyTokens[0], channel),
+        new DummyAccount(dummyTokens[1], channel)
+    ]
+    const tester = new TestBot(process.env.DISCORD_TEST_BOT_TOKEN, client, channel, ...dummyBots)
     await tester.init()
-    // const dummy1 = new TestBot()
     
-    await require('./commands.test.js')(client, tester)
-    await require('./api.test.js')(client, tester)
-    await require('./games.test.js')(client, tester)
+    await apiTest(client, tester)
+    await commandTest(client, tester)
+    await gameTest(client, tester)
 }
